extract influence explain panel component
